Basic graphics buffer and image draw working
